Redefine and restructure the repository
The idea is to use the same repository to maintain code developed when
playing with embedded systems.

The embedded programming from scratch application is moved into it's own
directory.

The repository name should be changed accordingly on GitHub. This avoids
crating more new repositories and also allows to maintain code developed
consistently.